api/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
distlib==0.3.0
distro==1.4.0
ecdsa==0.14.1
fastapi==0.61.2
google==3.0.0
google-auth==1.23.0
h11==0.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
loguru==0.5.3
msgpack==0.6.2
oauth==1.0.1
packaging==20.3
passlib==1.7.4
peewee==3.14.0
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.2
pyparsing==2.4.6
python-jose==3.2.0
python-multipart==0.0.5
pytoml==0.1.21
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
rsa==4.6
six==1.14.0
soupsieve==2.0.1
starlette==0.13.6
urllib3==1.25.8
uvicorn==0.12.2
webencodings==0.5.1
